requirements.txt
streamlit
pandas
numpy
google-generativeai
google-api-core
python-dotenv
